Remove unneeded controllers
