+++ contracts/base-factory/src/contract.rs
@@ -1,10 +1,14 @@
 #[cfg(not(feature = "library"))]
 use cosmwasm_std::entry_point;
-use cosmwasm_std::{to_binary, Binary, Deps, DepsMut, Env, MessageInfo, StdResult, WasmMsg};
+use cosmwasm_std::{
+    ensure_eq, to_binary, Binary, Deps, DepsMut, Env, MessageInfo, StdResult, WasmMsg,
+};
 use cw2::set_contract_version;
 use cw_utils::must_pay;
 use sg1::checked_fair_burn;
+use sg2::msg::UpdateMinterParamsMsg;
 use sg2::query::Sg2QueryMsg;
+use sg2::MinterParams;
 use sg_std::NATIVE_DENOM;
 
 use crate::error::ContractError;
@@ -13,7 +17,6 @@ use crate::msg::{
     ParamsResponse, Response, SudoMsg,
 };
 use crate::state::SUDO_PARAMS;
-use sg_controllers::update_params;
 
 // version info for migration info
 const CONTRACT_NAME: &str = "crates.io:sg-base-factory";
@@ -92,6 +95,36 @@ pub fn sudo_update_params(
     SUDO_PARAMS.save(deps.storage, &params)?;
 
     Ok(Response::new().add_attribute("action", "sudo_update_params"))
+}
+
+/// Base update params that can be used by other minter factories
+pub fn update_params<T, C>(
+    params: &mut MinterParams<C>,
+    param_msg: UpdateMinterParamsMsg<T>,
+) -> Result<(), ContractError> {
+    params.code_id = param_msg.code_id.unwrap_or(params.code_id);
+
+    if let Some(creation_fee) = param_msg.creation_fee {
+        ensure_eq!(
+            &creation_fee.denom,
+            &NATIVE_DENOM,
+            ContractError::InvalidDenom {}
+        );
+        params.creation_fee = creation_fee;
+    }
+
+    if let Some(min_mint_price) = param_msg.min_mint_price {
+        ensure_eq!(
+            &min_mint_price.denom,
+            &NATIVE_DENOM,
+            ContractError::InvalidDenom {}
+        );
+        params.min_mint_price = min_mint_price;
+    }
+
+    params.mint_fee_bps = param_msg.mint_fee_bps.unwrap_or(params.mint_fee_bps);
+
+    Ok(())
 }
 
 #[cfg_attr(not(feature = "library"), entry_point)]
