+++ contracts/base-factory/src/error.rs
@@ -1,7 +1,6 @@
 use cosmwasm_std::StdError;
 use cw_utils::PaymentError;
 use sg1::FeeError;
-use sg_controllers::MinterFactoryError;
 use thiserror::Error;
 
 #[derive(Error, Debug, PartialEq)]
@@ -15,12 +14,9 @@ pub enum ContractError {
     #[error("{0}")]
     Payment(#[from] PaymentError),
 
-    #[error("{0}")]
-    MinterFactory(#[from] MinterFactoryError),
-
     #[error("Unauthorized")]
     Unauthorized {},
 
-    #[error("MinterFactoryError")]
-    MinterFactoryError {},
+    #[error("InvalidDenom")]
+    InvalidDenom {},
 }
