+++ contracts/vending-factory/src/error.rs
@@ -1,7 +1,7 @@
+use base_factory::ContractError as BaseContractError;
 use cosmwasm_std::StdError;
 use cw_utils::PaymentError;
 use sg1::FeeError;
-use sg_controllers::MinterFactoryError;
 use thiserror::Error;
 
 #[derive(Error, Debug, PartialEq)]
@@ -15,14 +15,8 @@ pub enum ContractError {
     #[error("{0}")]
     Payment(#[from] PaymentError),
 
-    #[error("{0}")]
-    MinterFactory(#[from] MinterFactoryError),
-
     #[error("Unauthorized")]
     Unauthorized {},
-
-    #[error("MinterFactoryError")]
-    MinterFactoryError {},
 
     #[error("InvalidNumTokens {max}, min: 1")]
     InvalidNumTokens { max: u32, min: u32 },
@@ -30,9 +24,9 @@ pub enum ContractError {
     #[error("Invalid minting limit per address. max: {max}, min: 1, got: {got}")]
     InvalidPerAddressLimit { max: u32, min: u32, got: u32 },
 
-    #[error("InvalidDenom")]
-    InvalidDenom {},
-
     #[error("Minimum network mint price {expected} got {got}")]
     InsufficientMintPrice { expected: u128, got: u128 },
+
+    #[error("{0}")]
+    BaseError(#[from] BaseContractError),
 }
