+++ packages/controllers/src/lib.rs
@@ -1,5 +1,3 @@
 mod hooks;
-mod minter_factory;
 
 pub use hooks::{HookError, Hooks, HooksResponse};
-pub use minter_factory::{update_params, MinterFactoryError};
